chore(bigquery): refactor the interfaces and implementations namespaces
