+++ Doppler.BigQueryMicroservice/Repository/IUserRepository.cs
@@ -1,7 +1,7 @@
 using Doppler.BigQueryMicroservice.Entitites;
 using System.Threading.Tasks;
 
-namespace Doppler.BigQueryMicroservice.Repository.Interfaces
+namespace Doppler.BigQueryMicroservice.Repository
 {
     /// <summary>
     /// Extension of base contract
