+++ Doppler.BigQueryMicroservice/Repository/Implementation/BaseRepository.cs
@@ -1,14 +1,11 @@
 using Dapper;
 using Doppler.BigQueryMicroservice.Infrastructure;
-using Doppler.BigQueryMicroservice.Repository.Interfaces;
 using Microsoft.Extensions.Logging;
 using System;
-using System.Collections.Generic;
 using System.Data;
-using System.Linq;
 using System.Threading.Tasks;
 
-namespace Doppler.BigQueryMicroservice.Repository.Classes
+namespace Doppler.BigQueryMicroservice.Repository.Implementation
 {
     /// <summary>
     /// Base repository abstraction
